fix: Changed default export on ens card and moved mock fixtures
