+++ src/pages/Domains/index.tsx
@@ -1,7 +1,7 @@
 import { Flex } from 'rebass';
 import styled from 'styled-components';
-import EnsNameCard from '../../components/EnsNameCard';
-import { MockENSData } from '../../constants/mock_fixtures';
+import { ENSNameCard } from '../../components/EnsNameCard';
+import { MockENSData } from '../../mock_fixtures';
 
 const StyledDomainsWrapper = styled(Flex)`
   flex-wrap: wrap;
@@ -12,7 +12,7 @@ export function Domains() {
   return (
     <StyledDomainsWrapper>
       {MockENSData.map(({ ensDomain, imageUrl }) => {
-        return <EnsNameCard ensDomain={ensDomain} imageUrl={imageUrl} />;
+        return <ENSNameCard ensDomain={ensDomain} imageUrl={imageUrl} />;
       })}
     </StyledDomainsWrapper>
   );
